+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/Application.java
@@ -3,7 +3,7 @@ package ch.admin.bit.jeap.errorhandling;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.UserDetailsServiceAutoConfiguration;
+import org.springframework.boot.security.autoconfigure.UserDetailsServiceAutoConfiguration;
 
 @SpringBootApplication(exclude = UserDetailsServiceAutoConfiguration.class)
 @Slf4j