+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/infrastructure/persistence/OriginalTraceContext.java
@@ -28,4 +28,6 @@ public class OriginalTraceContext {
 
     private String traceIdString;
 
+    private Boolean sampled;
+
 }