+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/web/WebSecurityConfig.java
@@ -5,8 +5,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.security.web.servlet.util.matcher.PathPatternRequestMatcher;
 import org.springframework.security.web.util.matcher.NegatedRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
@@ -16,23 +17,22 @@ public class WebSecurityConfig  {
 
     @Bean
     @Order(Ordered.HIGHEST_PRECEDENCE + 11)
-    public SecurityFilterChain uiSecurityFilterChain(HttpSecurity http) throws Exception {
+    public SecurityFilterChain uiSecurityFilterChain(HttpSecurity http) {
         // csrf is unnecessary as we're already using an oauth token
         http.csrf(csrf -> csrf.disable());
 
         // protect the API
         // allow public access to frontend resources (i.e. non-/api-routes)
         // permit open access to open API docs & swagger ui as they are only enabled on test environments
-        RequestMatcher antPathMatcher = new AntPathRequestMatcher("/api/**");
-        RequestMatcher matcher = new NegatedRequestMatcher(antPathMatcher);
+        RequestMatcher apiMatcher = PathPatternRequestMatcher.withDefaults().matcher("/api/**");
+        RequestMatcher matcher = new NegatedRequestMatcher(apiMatcher);
         http.securityMatcher(matcher)
                 .authorizeHttpRequests(authorizeHttpRequests ->
                         authorizeHttpRequests.anyRequest().permitAll());
 
         // this is used for the auth - silent-renew.html
         http.headers(headers ->
-                headers.frameOptions( frameOptions ->
-                        frameOptions.sameOrigin()));
+                headers.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin));
 
         return http.build();
     }