+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/web/ui/configuration/ConfigurationControllerSecurityConfig.java
@@ -7,7 +7,7 @@ import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.security.web.servlet.util.matcher.PathPatternRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 
 /**
@@ -16,12 +16,12 @@ import org.springframework.security.web.util.matcher.RequestMatcher;
 @Configuration
 class ConfigurationControllerSecurityConfig {
     private static RequestMatcher configurationServiceMatcher() {
-        return new AntPathRequestMatcher("/api/configuration/**", HttpMethod.GET.name());
+        return PathPatternRequestMatcher.withDefaults().matcher(HttpMethod.GET, "/api/configuration/**");
     }
 
     @Bean
     @Order(Ordered.HIGHEST_PRECEDENCE + 12)
-    SecurityFilterChain configSecurityFilterChain(HttpSecurity http) throws Exception {
+    SecurityFilterChain configSecurityFilterChain(HttpSecurity http) {
         http.securityMatcher(configurationServiceMatcher());
         http.authorizeHttpRequests( authorizeHttpRequests ->
                 authorizeHttpRequests.anyRequest().permitAll());