+++ jeap-error-handling-service/src/test/java/ch/admin/bit/jeap/errorhandling/infrastructure/manualtask/TaskManagementClientTest.java
@@ -17,7 +17,7 @@ import ch.admin.bit.jeap.security.test.jws.JwsBuilderFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.client.RestClientTest;
+import org.springframework.boot.restclient.test.autoconfigure.RestClientTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.client.HttpClientErrorException;