+++ jeap-error-handling-service/src/test/java/ch/admin/bit/jeap/errorhandling/infrastructure/persistence/PersistenceTestConfig.java
@@ -1,6 +1,6 @@
 package ch.admin.bit.jeap.errorhandling.infrastructure.persistence;
 
-import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
+import org.springframework.boot.flyway.autoconfigure.FlywayMigrationStrategy;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;